[#41] refactor: 드롭다운에 jsx 담을 수 있게 수정
